# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Students Week 2026 food ordering API
servers:
  - url: /api
...
paths:
...
  /side-items:
    post:
      operationId: createSideItem
      tags: [menu]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SideItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideItem"
  /side-items/{id}:
    patch:
      operationId: updateSideItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SideItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SideItem"
    delete:
      operationId: deleteSideItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /protein-items:
    post:
      operationId: createProteinItem
      tags: [menu]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProteinItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProteinItem"
  /protein-items/{id}:
    patch:
      operationId: updateProteinItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProteinItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProteinItem"
    delete:
      operationId: deleteProteinItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, served]
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
          description: Search by customer name or table number
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderFull"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFull"
        "409":
          description: Duplicate order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/history:
    get:
      operationId: getOrderHistory
      tags: [orders]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, served]
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Order history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderFull"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFull"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderFull"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SideItem:
      type: object
      required: [id, mainDishTypeId, name, isComplementary, isAvailable]
      properties:
        id:
          type: integer
        mainDishTypeId:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        isComplementary:
          type: boolean
        isAvailable:
          type: boolean
    SideItemInput:
      type: object
      required: [mainDishTypeId, name]
      properties:
        mainDishTypeId:
          type: integer
        name:
          type: string
          minLength: 1
        imageUrl:
          type: string
        isComplementary:
          type: boolean
        isAvailable:
          type: boolean
    SideItemUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
        isComplementary:
          type: boolean
        isAvailable:
          type: boolean
    ProteinItem:
      type: object
      required: [id, mainDishTypeId, name, isAvailable]
      properties:
        id:
          type: integer
        mainDishTypeId:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
    ProteinItemInput:
      type: object
      required: [mainDishTypeId, name]
      properties:
        mainDishTypeId:
          type: integer
        name:
          type: string
          minLength: 1
        imageUrl:
          type: string
        isAvailable:
          type: boolean
    ProteinItemUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
...
    OrderSideItem:
      type: object
      required: [id, sideId, sideName, isComplementary]
      properties:
        id:
          type: integer
        sideId:
          type: integer
        sideName:
          type: string
        isComplementary:
          type: boolean
    OrderProteinItem:
      type: object
      required: [id, proteinId, proteinName]
      properties:
        id:
          type: integer
        proteinId:
          type: integer
        proteinName:
          type: string
    OrderFull:
      type: object
      required: [id, customerName, tableNumber, vendorId, vendorName, status, createdAt, sides, proteins]
      properties:
        id:
          type: integer
        customerName:
          type: string
        tableNumber:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        mainDishTypeId:
          type: ["integer", "null"]
        mainDishTypeName:
          type: ["string", "null"]
        mainDishName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, served]
        createdAt:
          type: string
        sides:
          type: array
          items:
            $ref: "#/components/schemas/OrderSideItem"
        proteins:
          type: array
          items:
            $ref: "#/components/schemas/OrderProteinItem"
    OrderInput:
      type: object
      required: [customerName, tableNumber, vendorId, mainDishTypeId, sideIds, proteinIds]
      properties:
        customerName:
          type: string
          minLength: 1
        tableNumber:
          type: string
          minLength: 1
        vendorId:
          type: integer
        mainDishTypeId:
          type: integer
        sideIds:
          type: array
          items:
            type: integer
        proteinIds:
          type: array
          items:
            type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, served]
...