# lib/api-zod/src/generated/types/sideItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Students Week 2026 food ordering API
 * OpenAPI spec version: 0.3.0
 */

export interface SideItemInput {
  mainDishTypeId: number;
  /** @minLength 1 */
  name: string;
  imageUrl?: string;
  isComplementary?: boolean;
  isAvailable?: boolean;
}