# lib/api-zod/src/generated/types/sideItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Students Week 2026 food ordering API
 * OpenAPI spec version: 0.3.0
 */

export interface SideItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  imageUrl?: string | null;
  isComplementary?: boolean;
  isAvailable?: boolean;
}